prepare project structure by moving to pkg/
