+++ pkg/k8sclient/spec.go
@@ -9,7 +9,7 @@ import (
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/giantswarm/k8sclient/k8scrdclient"
+	"github.com/giantswarm/k8sclient/pkg/k8scrdclient"
 )
 
 type Interface interface {
