+++ pkg/k8sclienttest/k8sclienttest.go
@@ -10,7 +10,7 @@ import (
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/giantswarm/k8sclient/k8scrdclient"
+	"github.com/giantswarm/k8sclient/pkg/k8scrdclient"
 )
 
 type ClientsConfig struct {
