+++ pkg/k8scrdclient/k8scrdclient.go
@@ -12,7 +12,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/giantswarm/k8sclient/k8sversion"
+	"github.com/giantswarm/k8sclient/pkg/k8sversion"
 )
 
 type Config struct {
